+++ src/config/config.hh
@@ -65,7 +65,7 @@ namespace http
 
     struct TimeoutConfig
     {
-        TimeoutConfig(float to_ka, float to_tran, float to_thr_val, float to_thr_time);
+        TimeoutConfig(double to_ka, double to_tran, double to_thr_val, double to_thr_time);
         TimeoutConfig() = default;
         TimeoutConfig(const TimeoutConfig&) = default;
         TimeoutConfig& operator=(const TimeoutConfig&) = default;
@@ -74,10 +74,10 @@ namespace http
 
         ~TimeoutConfig() = default;
 
-        float to_keep_alive_;
-        float to_transaction_;
-        float to_throughput_val_;
-        float to_throughput_time_;
+        double to_keep_alive_;
+        double to_transaction_;
+        double to_throughput_val_;
+        double to_throughput_time_;
     };
     struct TimeoutConfig set_TimeoutConfig(const std::string& str);
     json get_timeouts(const std::string& s);
